[ADD]Adjust code style,Added delete leave people func
