+++ cms/templates/maintenance/index.html
@@ -18,6 +18,7 @@ from django.urls import reverse
     </ul>
     <div class="dingtalk-bottom-box">
         <a href="#" class="dingtalk-bottom">${_('Import users from dingtalk')}</a>
+        <a href="#" class="dingtalk-button-delete">${_('Delete terminated employee')}</a>
     </div>
     <div class="dingtalk-user-info-box is-shown">
         <p>${_("Success user info:")}</p>
@@ -25,6 +26,11 @@ from django.urls import reverse
         </ul>
         <p>${_("fail user info:")}</p>
         <ul class="dingtalk-user-info-list" id="fail_add_user_info">
+        </ul>
+    </div>
+    <div class="dingtalk-delete-result result-is-shown">
+        <p>${_('Delete result:')}&nbsp;&nbsp;&nbsp;&nbsp;<span id="dingtalk-delete-msg"></span></p>
+        <ul class="dingtalk-delete-user-detail">
         </ul>
     </div>
 </div>
